Add Hashing negative result case unit test & Move FileManagerService to sub-directory
