+++ CommonServices/FileServices/FileManager/IFileManagerService.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.IO;
 using System.Text;
 
-namespace CommonServices
+namespace CommonServices.FileServices.FileManager
 {
     public interface IFileManagerService
     {
