+++ CommonServicesTests/FileManagerServiceTests.cs
@@ -1,9 +1,13 @@
-﻿using System;
-using System.IO;
-using CommonServices;
+﻿using CommonServices;
 using CommonServices.DatabaseOperator;
+using CommonServices.FileServices.FileManager;
+using Microsoft.Extensions.Configuration;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using Moq;
+using System;
+using System.Collections.Generic;
+using System.IO;
+using System.Text;
 
 namespace CommonServicesTests
 {
@@ -12,6 +16,7 @@ namespace CommonServicesTests
     {
         const string RESOURCES_DIRECTORY = "resources";
         const string TEST_FILE = "testFile.txt";
+        const string MODIFIED_TEST_FILE = "testFile2.txt";
         const string EXPECTED_TEST_FILE_SHA256_CHECKSUM = "1F0D0DFE1F134E4FDEB01A0627B127CE6BAC8D171220B0A9EF14D56E755A460A";        
         [TestMethod]
         public void Can_Get_Correct_SHA256_Checksum()
@@ -21,10 +26,30 @@ namespace CommonServicesTests
             FileManagerService objectUnderTest = new FileManagerService(mockConfiguration, mockDatabaseConnector.Object);
 
             string workingDirectory = Path.GetFullPath(Directory.GetCurrentDirectory());
-            Console.WriteLine(workingDirectory);
-            var result = objectUnderTest.ComputeFileSHA256Checksum(workingDirectory + Path.DirectorySeparatorChar + RESOURCES_DIRECTORY + Path.DirectorySeparatorChar + TEST_FILE);
+            string testFilePath = workingDirectory + Path.DirectorySeparatorChar + RESOURCES_DIRECTORY + Path.DirectorySeparatorChar + TEST_FILE;
+            var result = objectUnderTest.ComputeFileSHA256Checksum(testFilePath);
 
+            Assert.IsNotNull(result);
             Assert.AreEqual(EXPECTED_TEST_FILE_SHA256_CHECKSUM, result);
+        }
+
+        [TestMethod]
+        public void Modified_File_Has_Incorrect_Checksum()
+        {
+            var mockConfiguration = CommonMethods.CreateMockConfiguration();
+            var mockDatabaseConnector = new Mock<IDatabaseConnector>();
+            FileManagerService objectUnderTest = new FileManagerService(mockConfiguration, mockDatabaseConnector.Object);
+
+            string workingDirectory = Path.GetFullPath(Directory.GetCurrentDirectory());
+            string testFilePath = workingDirectory + Path.DirectorySeparatorChar + RESOURCES_DIRECTORY + Path.DirectorySeparatorChar + TEST_FILE;
+            string modifiedTestFilePath = workingDirectory + Path.DirectorySeparatorChar + RESOURCES_DIRECTORY + Path.DirectorySeparatorChar + MODIFIED_TEST_FILE;
+            
+            var testFileResult = objectUnderTest.ComputeFileSHA256Checksum(testFilePath);
+            var modifiedTestFileResult = objectUnderTest.ComputeFileSHA256Checksum(modifiedTestFilePath);
+            
+            Assert.IsNotNull(testFileResult);
+            Assert.IsNotNull(modifiedTestFileResult);
+            Assert.AreNotEqual(modifiedTestFileResult, testFileResult);
         }
     }
 }
