+++ CommonServicesTests/resources/testFile2.txt
@@ -1,0 +1,1 @@
+﻿Modified test file
